requirements.txt
Flask==1.1.1
PyYAML==5.3
flask-swagger==0.2.14
xmltodict==0.12.0
Flask-HTTPAuth==3.3.0
